requirements.txt
black==22.10.0
click==8.1.3
coverage==6.5.0
coveralls==3.3.1
Django==2.2
django-crispy-forms==1.14.0
django-filter==2.3.0
Pillow==9.2.0
flake8==5.0.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0

